basis.router: split parametrized route tests into several files
